Object hierarchy is now restored upon import.
The algorithm isn't very efficient as it loops, for each object, over
all the objects in the scene.

A more straightforward way would be to build a tree from the Alembic
archive and use this for the actual object hierarchy, however every
attempt at this were unsuccessful. This is to be revisited in the
future.