Revert changes done to the transform matrix read/write and axis conversion code to the state of the original patch.
The rewrite wasn't ideal, and prone to bugs. The idea was to not
hardcode Y-up <-> Z-up conversions but rather let the user choose the
axis during the export/import process, but since the other DCCs commonly
found in VFX pipeline all use a Y-up system, hardcoding everything
should be fine for the time being. This will probably be revisited in
the future if need be.