Improve (fix a few bugs in) the transform hierarchy creation.
Now an empty is created for every transform node except for the cases
when the transform node only has a single child that is not a transform
itself.

There are still a few issues in complex hierarchies where multiple
objects have the same name (which is legal in Alembic since it is the
whole path to the object that counts, not just the name).